#增加了GrapX vs2017工程  -去掉了DX9的支持  *修改了C++14下   *union没有显式构造函数报错的问题   *宏和字符串连写没有空格会报错的问题   *union赋值编译不通过的问题  +nuGet增加了DirectXTex库

+++ GrapX/Platform/Win32_D3D11/GXGraphicsImpl_d3d11.cpp
@@ -691,8 +691,8 @@ namespace D3D11
       IntGetDimension((GXUINT&)width, (GXUINT&)height);
       Viewport.TopLeftX = 0;
       Viewport.TopLeftY = 0;
-      Viewport.Width    = width;
-      Viewport.Height   = height;
+      Viewport.Width    = static_cast<FLOAT>(width);
+      Viewport.Height   = static_cast<FLOAT>(height);
       Viewport.MinDepth = 0.f;
       Viewport.MaxDepth = 1.f;
     }
